implement favourite and bookmark in gui. It works but needs more testing. fixed: now catching exceptions on rateShot.
Because of the these two buttons are needing four states (disabled, hover, active, inactive) it was not possible to use the textures. There is no colordiffuse or similar on buttons.
I implemented it by changing the textcolor depending on the state active, inactive and disabled. The hover state is defined in the skin.

If you see a better way (except putting imgs over the buttons...) please revert mine ;)