added offline_debug in api.
Allow using a given OFFLINE_SHOT (dict) to be used instead asking the server. This is useful for testing (e.g. gui) without stressing the server.
To aktive this set OFFLINE_DEBUG to True in api and alter the OFFLINE_SHOT dict. Please don't commit api with OFFLINE_DEBUG set to True!

This also documents all values in shot dict.

Maybe this will be reverted or branched before the (beta) official release.