add feature: configurable hotkey assigning
You can now configure (in the xbmc settings) hotkeys for the 9 main functions. You can choose one of 18 keys.
Not all of these keys are tested - please report if some aren't working!

I changed the hotkey for jump from context to info-key because jump is more used for debugging - if we (beta-)release we should de-assign the info key.
The context-key is now default assigned to guess shot - which will be needed more often.

fixed: solveShot was not possible if shot was unsolvable because the button was disabled. Now with the hotkey funtion is was needed to real check if the actual shot is solvable before calling the api function.