added fifth shot type 'The Vault', rewrote the date part in gui and api
there is a fifth category of shots "The Vault", example: 94296
This shots dont have a date and will never have a date - this required a rewrote of the date section in the api.
I also improved the type of date by using "datetime.date" instead of "datetime.datetime" because there never will be a time.

thanks to fuxii for finding this bug :)