added multi threading shot preloading
There are now 3 background scraper workers which will each wait for a free scrape-job. 3 'random' jobs are automatically placed in the queue of jobs. If a shot is requested its job is placed in the queue of jobs. Next step is checking if a shot of this type is already present in the list of preloaded shots - if not it will wait for 1 second and check again. If a matching shot is found it is given to the gui.

The game is now really fast :)